python/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pyserial==3.5
pydantic==2.9.2
pydantic-settings==2.5.2
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.3
jinja2==3.1.4